+++ ios/Podfile
@@ -51,5 +51,9 @@ end
 post_install do |installer|
   installer.pods_project.targets.each do |target|
     flutter_additional_ios_build_settings(target)
+    target.build_configurations.each do |config|
+      definitions = config.build_settings['GCC_PREPROCESSOR_DEFINITIONS'] ||= ['$(inherited)']
+      definitions << 'PERMISSION_MEDIA_LIBRARY=1'
+    end
   end
 end
