+++ pubspec.yaml
@@ -25,10 +25,10 @@ dependencies:
   shared_preferences: ^2.3.3
   path_provider: ^2.1.5
   percent_indicator: ^4.2.5
-  firebase_core: ^3.15.2
-  firebase_crashlytics: ^4.3.10
-  firebase_analytics: ^11.6.0
-  cloud_functions: ^5.6.2
+  firebase_core: ^3.12.1
+  firebase_crashlytics: ^4.3.4
+  firebase_analytics: ">=11.4.4 <11.5.2"
+  cloud_functions: 5.3.4
   package_info_plus: ^9.0.1
   url_launcher: ^6.3.2
   lottie: ^3.1.3
